[red-knot] refactor CFG outside of symbol table
