+++ crates/red_knot/src/program/mod.rs
@@ -43,7 +43,7 @@ impl Program {
         let (source, semantic, lint) = self.jars_mut();
         for change in aggregated_changes.iter() {
             semantic.module_resolver.remove_module_by_file(change.id);
-            semantic.symbol_tables.remove(&change.id);
+            semantic.semantic_indices.remove(&change.id);
             source.sources.remove(&change.id);
             source.parsed.remove(&change.id);
             // TODO: remove all dependent modules as well
